na.py: initial commit
Add a simple Python script that can take in a YAML bool as arg 1 and
echo arg 2 back to stdout. This is useful for workflows where a caller
specifies acboolean to add an argument to the script. This approach
was used over a raw extra options string to prevent users from having
to know the interface into build.sh.

Keep the script name short to keep the using command lines short.

Signed-off-by: William Roberts <william.c.roberts@intel.com>