ci: add an action suitable for pull requests
The current set of actions build and upload images as test
and then attempt to use them for rendering.  Not only is their
a race condition for submitted PRs, it also requiresd PRs to have
permissions to push into the org packages. We really just
want to build the dockerfile and attempt to render the
markdown files.

This also adds a test script for running CI tests locally for debugging.

Signed-off-by: William Roberts <william.c.roberts@intel.com>