blow away the python filter for informative text; make a lua filter later if needed (#91)
This change takes 200MB out of the image by removing the one py3 filter we had (which was a simple filter to apply the informative text style for Word export). We can add this back as a lua filter if it's truly needed, based on all the other types of text blocks.