(40x speedup) Rewrite the tabularray extension to use xltabular (#112)
tabularray is beautiful and easy to work with, but has some serious performance drawbacks. This change rewrites the extension by using xltabular (which is an addon to tabularx to support longtable-style continuation).

Performance testing indicate this change gives about a 40x speedup on top of the previous fixes to use XeLaTeX and only two passes (tested on my arm64 laptop):

Before this change, rendering Part 2 of the TPM spec (from TeX) takes 520 seconds.
After this change, rendering Part 2 of the TPM spec (from TeX) takes 13 seconds.
This change reverts some intermediate style changes to tables that were made to closely match the 'default' TCG table style (gray on gray on gray). tabularray made this easy (at a huge cost of performance). Turns out, tabularray was doing something very fancy to support colored cells across rowspans. The usual ways of coloring table cells (\rowcolors, \cellcolor) break for multirow situations. Also, (subjective) the gray-on-gray-on-gray style is distracting. Tables have been reverted to a clean black and white style for now, given that this was also always allowed. A future feature-request and pull-request to style the tables some more would be encouraged for anybody who really wants to see this style change.