Update dependencies
This change updates the dependency versions of:

- debian-bookworm
- pandoc
- pandoc-crossref
- puppeteer
- imgur
- mermaid-filter
- typescript
- drawio

It also resolves an issue seen running on newer Chromium browsers (>126
or so) where the browser hangs. See
https://github.com/puppeteer/puppeteer/issues/11640 for more information
about this issue.

Because of the update to pandoc 3.2.1, this change also updates the TCG
LaTeX template to provide the new required `\pandocbounded` macro. It
doesn't appear to have any actual effect for images in docs.