Add citations feature (+ backmatter). (#206)
Adds support for citations, and introduce `\beginbackmatter` command to
mark back matter sections. This avoids the references/bibliography to be
part of the appendices (which is wrong, references/bibliography are no
appendix).

You can write something like "The TPM 2.0 architecture
[@{cite:tcg-tpm2-architecture}] describes ..." in your Markdown document
which results in a
clickable link (e.g., "The TPM 2.0 architecture [1] describes ...") to
the references/citations list at the
end of the PDF document.

In the YAML front matter you add something like:

```yaml
---
bibliography: citations.yaml
csl: transactions-on-computer-systems.csl
...
```

The `citations.yaml` may look as follows:

```yaml
references:
  - id: "cite:tcg-tpm2-architecture"
    title: "Trusted Platform Module Library - Part 1: Architecture"
    author:
      - family: "(TCG)"
        given: "Trusted Computing Group"
    edition: "Family 2.0, Level 00, Revision 01.83"
    organization: "Trusted Computing Group"
    publisher: "Trusted Computing Group (TCG)"
    series: "TCG Specification"
    issued:
      year: 2024
      month: 1
      day: 25
url:
"https://trustedcomputinggroup.org/resourcetpm-library-specification/"
```

For the back matter, please put the `\beginbackmatter` command in the
Markdown source and add the references as follows:

```md
\beginbackmatter

# References

::: {#refs}
:::
```

Signed-off-by: Michael Eckel <michael.eckel@sit.fraunhofer.de>