Enable (and fix) Pandoc syntax highlighting (#243)
This change removes the --no-highlight flag from the pandoc invocation.
It fixes the issue with the template that that revealed (which is that
the $highlighting-styles$ magic was missing). If you are using this
project with a custom LaTeX template and this change breaks with an
error like "Environment Shaded undefied", then it's because your
template needs this magic as well. See
https://github.com/jgm/pandoc/issues/7923 for more information on that.

This change re-styles code blocks to use a white background instead of
the light blue one, so that the kate-styled syntax highlighting looks
good. Since the background is white, but the code block still needs to
be a block, this change adds a 1pt black line around the code.