Allow all code blocks to span multiple pages.
Currently, code blocks with classes are rendered with the Shaded
environment, which disables page breaking.

This change redefines that environment to use tcolorbox, which does
allow page breaking. This environment can be used exclusively for code
blocks, allowing the customcodeblock environment to be retired.

This change also adds a dummy class to all code blocks, ensuring they
all use the Shaded environment and pick up the tcolorbox tweaks.